requirements.txt
diffusers
transformers
accelerate
segmentation-models-pytorch
opencv-python
xformers
torch
Pillow
PyYAML